refactor: OrderTable, TableGroup package 분리
