+++ src/main/java/kitchenpos/order/domain/value/OrderTables.java
@@ -4,8 +4,8 @@ import java.util.Collections;
 import java.util.List;
 import javax.persistence.Embeddable;
 import javax.persistence.OneToMany;
-import kitchenpos.order.domain.entity.OrderTable;
-import kitchenpos.order.domain.entity.TableGroup;
+import kitchenpos.table.domain.entity.OrderTable;
+import kitchenpos.table.domain.entity.TableGroup;
 
 @Embeddable
 public class OrderTables {
