+++ src/main/java/kitchenpos/order/service/OrderService.java
@@ -9,16 +9,16 @@ import kitchenpos.menu.exception.NotFoundMenuException;
 import kitchenpos.order.domain.entity.Order;
 import kitchenpos.order.domain.entity.OrderLineItem;
 import kitchenpos.order.domain.entity.OrderRepository;
-import kitchenpos.order.domain.entity.OrderTable;
-import kitchenpos.order.domain.entity.OrderTableRepository;
+import kitchenpos.table.domain.entity.OrderTable;
+import kitchenpos.table.domain.entity.OrderTableRepository;
 import kitchenpos.order.domain.value.OrderLineItems;
 import kitchenpos.order.domain.value.Quantity;
 import kitchenpos.order.dto.OrderRequest;
 import kitchenpos.order.dto.OrderResponse;
 import kitchenpos.order.exception.NotFoundOrderException;
-import kitchenpos.order.exception.NotFoundOrderTableException;
+import kitchenpos.table.exception.NotFoundOrderTableException;
 import kitchenpos.order.exception.OrderLineItemIsNullOrZeroException;
-import kitchenpos.order.exception.OrderTableIsEmptyException;
+import kitchenpos.table.exception.OrderTableIsEmptyException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
