+++ src/main/java/kitchenpos/table/exception/OrderTableCountException.java
@@ -1,4 +1,4 @@
-package kitchenpos.order.exception;
+package kitchenpos.table.exception;
 
 public class OrderTableCountException extends IllegalArgumentException {
