+++ src/main/java/kitchenpos/table/exception/OrderTableIsNotEmptyException.java
@@ -1,4 +1,4 @@
-package kitchenpos.order.exception;
+package kitchenpos.table.exception;
 
 public class OrderTableIsNotEmptyException extends IllegalArgumentException {
