+++ src/test/java/kitchenpos/order/ui/OrderRestControllerTest.java
@@ -13,7 +13,7 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import java.nio.charset.StandardCharsets;
 import java.util.Arrays;
-import kitchenpos.order.domain.entity.OrderTable;
+import kitchenpos.table.domain.entity.OrderTable;
 import kitchenpos.order.domain.value.OrderStatus;
 import kitchenpos.order.dto.OrderLineItemRequest;
 import kitchenpos.order.dto.OrderRequest;
