+++ src/test/java/kitchenpos/table/service/TableGroupService2Test.java
@@ -1,44 +1,28 @@
-package kitchenpos.order.service;
+package kitchenpos.table.service;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertAll;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.ArgumentMatchers.anyList;
-import static org.mockito.ArgumentMatchers.anyLong;
-import static org.mockito.Mockito.when;
 
 import java.util.Arrays;
-import java.util.Optional;
-import kitchenpos.order.domain.entity.OrderRepository;
-import kitchenpos.order.domain.entity.OrderTable;
-import kitchenpos.order.domain.entity.OrderTableRepository;
-import kitchenpos.order.domain.entity.TableGroup;
-import kitchenpos.order.domain.entity.TableGroupRepository;
-import kitchenpos.order.domain.value.NumberOfGuests;
+import kitchenpos.table.domain.entity.OrderTable;
+import kitchenpos.table.domain.entity.TableGroup;
+import kitchenpos.table.domain.value.NumberOfGuests;
 import kitchenpos.order.domain.value.OrderTables;
-import kitchenpos.order.dto.OrderTableRequest;
-import kitchenpos.order.dto.TableGroupRequest;
-import kitchenpos.order.dto.TableGroupResponse;
+import kitchenpos.table.dto.OrderTableRequest;
+import kitchenpos.table.dto.TableGroupRequest;
+import kitchenpos.table.dto.TableGroupResponse;
+import kitchenpos.table.service.TableGroupService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.InjectMocks;
-import org.mockito.Mock;
-import org.mockito.junit.jupiter.MockitoExtension;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
 
-@ExtendWith(MockitoExtension.class)
-class TableGroupServiceTest {
+@SpringBootTest
+class TableGroupService2Test {
 
-    @Mock
-    OrderRepository orderRepository;
-    @Mock
-    OrderTableRepository orderTableRepository;
-    @Mock
-    TableGroupRepository tableGroupRepository;
-
-    @InjectMocks
+    @Autowired
     TableGroupService tableGroupService;
 
     OrderTable 오더테이블_테이블1;
@@ -53,9 +37,9 @@ class TableGroupServiceTest {
     @BeforeEach
     void setUp() {
         오더테이블_테이블1 = new OrderTable(1L, null, NumberOfGuests.of(44), true);
-        오더테이블_테이블1_리퀘스트 = new OrderTableRequest(1L, null, 3, true);
+        오더테이블_테이블1_리퀘스트 = new OrderTableRequest(11L, null, 3, true);
         오더테이블_테이블2 = new OrderTable(2L, null, NumberOfGuests.of(33), true);
-        오더테이블_테이블2_리퀘스트 = new OrderTableRequest(2L, null, 3, true);
+        오더테이블_테이블2_리퀘스트 = new OrderTableRequest(12L, null, 3, true);
 
         테이블그룹_테이블1_테이블2 = new TableGroup(1L,
             new OrderTables(Arrays.asList(오더테이블_테이블1, 오더테이블_테이블2)));
@@ -68,16 +52,11 @@ class TableGroupServiceTest {
     @Test
     @DisplayName("단체 테이블을 생성한다.")
     void create() {
-        //given
-        when(orderTableRepository.findAllById(anyList()))
-            .thenReturn(Arrays.asList(오더테이블_테이블1, 오더테이블_테이블2));
-        when(tableGroupRepository.save(any())).thenReturn(테이블그룹_테이블1_테이블2);
-
         //when
         TableGroupResponse createdTableGroup = tableGroupService.create(테이블그룹_테이블1_테이블2_리퀘스트);
 
         //then
-        assertThat(createdTableGroup.getId()).isEqualTo(테이블그룹_테이블1_테이블2.getId());
+        assertThat(createdTableGroup).isNotNull();
     }
 
     @Test
@@ -95,10 +74,6 @@ class TableGroupServiceTest {
         TableGroupRequest 테이블그룹_테이블1_테이블1_리퀘스트 = new TableGroupRequest(
             Arrays.asList(오더테이블_테이블1_리퀘스트, 오더테이블_테이블1_리퀘스트));
 
-        when(
-            orderTableRepository.findAllById(Arrays.asList(오더테이블_테이블1.getId(), 오더테이블_테이블1.getId())))
-            .thenReturn(Arrays.asList(오더테이블_테이블1));
-
         //when && then
         assertThatThrownBy(() -> tableGroupService.create(테이블그룹_테이블1_테이블1_리퀘스트))
             .isInstanceOf(IllegalArgumentException.class);
@@ -110,10 +85,6 @@ class TableGroupServiceTest {
         //given
         오더테이블_테이블2_리퀘스트 = new OrderTableRequest(2L, null, 3, false);
         오더테이블_테이블2 = new OrderTable(2L, NumberOfGuests.of(4), false);
-        when(
-            orderTableRepository.findAllById(Arrays.asList(오더테이블_테이블1.getId(), 오더테이블_테이블2.getId())))
-            .thenReturn(Arrays.asList(오더테이블_테이블1, 오더테이블_테이블2));
-        //when(tableGroupRepository.save(any())).thenReturn(테이블그룹_테이블1_테이블2);
 
         //when && then
         assertThatThrownBy(() -> tableGroupService.create(테이블그룹_테이블1_테이블2_리퀘스트))
@@ -127,9 +98,6 @@ class TableGroupServiceTest {
         TableGroup tableGroup = new TableGroup();
         오더테이블_테이블2 = new OrderTable(2L, tableGroup, NumberOfGuests.of(4), true);
 
-        when(orderTableRepository.findAllById(anyList()))
-            .thenReturn(Arrays.asList(오더테이블_테이블1, 오더테이블_테이블2));
-
         //when && then
         assertThatThrownBy(() -> tableGroupService.create(테이블그룹_테이블1_테이블2_리퀘스트))
             .isInstanceOf(IllegalArgumentException.class);
@@ -138,16 +106,8 @@ class TableGroupServiceTest {
     @Test
     @DisplayName("단체테이블을 해체한다.")
     void ungroup() {
-        //given
-        when(tableGroupRepository.findById(테이블그룹_테이블1_테이블2.getId()))
-            .thenReturn(Optional.of(테이블그룹_테이블1_테이블2));
-        when(orderTableRepository.findAllByTableGroupId(anyLong()))
-            .thenReturn(Arrays.asList(오더테이블_테이블1, 오더테이블_테이블2));
-        when(orderRepository.existsByOrderTableIdInAndOrderStatusIn(any(), any()))
-            .thenReturn(false);
-
         //when
-        tableGroupService.ungroup(테이블그룹_테이블1_테이블2.getId());
+        tableGroupService.ungroup(99L);
 
         //then
         assertAll(() -> {
@@ -159,13 +119,6 @@ class TableGroupServiceTest {
     @Test
     @DisplayName("조리중이거나 식사중일 경우 해체에 실패한다.")
     void ungroup_with_exception_when_orderStatus_is_cooking_or_meal() {
-        //given
-        when(tableGroupRepository.findById(테이블그룹_테이블1_테이블2.getId()))
-            .thenReturn(Optional.of(테이블그룹_테이블1_테이블2));
-        when(orderTableRepository.findAllByTableGroupId(anyLong()))
-            .thenReturn(Arrays.asList(오더테이블_테이블1, 오더테이블_테이블2));
-        when(orderRepository.existsByOrderTableIdInAndOrderStatusIn(any(), any())).thenReturn(true);
-
         //when && then
         assertThatThrownBy(() -> tableGroupService.ungroup(테이블그룹_테이블1_테이블2.getId()))
             .isInstanceOf(IllegalArgumentException.class);
